Remove python39-{pip,setuptools,wheel}
It was still working, with both the python39-{pip,setuptools,wheel}
and the python-{pip,setuptools,wheel}-wheel packages. Let's see if
the latter are actually sufficient. I think they may be, since all
use of pip, setuptools, and wheel should be in the virtual
environment, where maybe `venv` can install the necessary PyPA
packages so long as their wheels are available on the system.